backend-python/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
face-recognition==1.3.0
numpy==1.21.2
opencv-python==4.5.3.56
pillow==8.3.2
pymongo==3.12.0
bcrypt==3.2.0
pyjwt==2.1.0
